Switch to a single crate that creates the binaries
