+++ Cargo.toml
@@ -7,6 +7,7 @@ members = [
     "ntp-os-clock",
     "ntp-udp",
     "test-binaries",
+    "ntpd"
 ]
 exclude = [ ]
