+++ ntp-metrics-exporter/src/lib.rs
@@ -1,270 +1,120 @@
-use std::sync::atomic::{AtomicU64, Ordering};
+#![forbid(unsafe_code)]
 
-use ntp_daemon::{observer::WrappedSocketAddr, ObservablePeerState, ObservableState};
-use ntp_os_clock::UnixNtpClock;
-use ntp_proto::NtpClock;
-use prometheus_client::{
-    encoding::EncodeLabelSet,
-    metrics::{counter::Counter, family::Family, gauge::Gauge},
-    registry::{Registry, Unit},
+mod metrics;
+
+pub use metrics::Metrics;
+use serde_json::json;
+use thiserror::Error;
+
+use std::{net::SocketAddr, path::PathBuf};
+
+use axum::{
+    http::{HeaderMap, HeaderValue, StatusCode},
+    response::{AppendHeaders, IntoResponse},
+    routing::get,
+    Json, Router,
 };
+use clap::Parser;
+use ntp_daemon::{Config, ObservableState};
 
-#[derive(Debug, Clone, PartialEq, Eq, Hash, EncodeLabelSet)]
-struct PeerLabels {
-    address: String,
+#[derive(Parser)]
+#[command(version = "0.2.0", about = "Serve ntpd-rs openmetrics via http")]
+struct Cli {
+    /// Which configuration file to read the socket paths from
+    #[arg(short, long)]
+    config: Option<PathBuf>,
+
+    /// Path of the observation socket
+    #[arg(short, long)]
+    observation_socket: Option<PathBuf>,
+
+    #[arg(short = 'l', long = "listen", default_value = "127.0.0.1:9975")]
+    listen_socket: SocketAddr,
 }
 
-#[derive(Debug, Clone, PartialEq, Eq, Hash, EncodeLabelSet)]
-struct ServerLabels {
-    listen_address: WrappedSocketAddr,
+#[derive(Debug, Error)]
+enum ServeError {
+    #[error("io error: {0}")]
+    Io(#[from] std::io::Error),
+    #[error("fmt error: {0}")]
+    Fmt(#[from] std::fmt::Error),
 }
 
-#[derive(Default)]
-pub struct Metrics {
-    system_poll_interval: Gauge<f64, AtomicU64>,
-    system_poll_interval_exp: Gauge,
-    system_precision: Gauge<f64, AtomicU64>,
-    system_accumulated_steps: Gauge<f64, AtomicU64>,
-    system_accumulated_steps_threshold: Gauge<f64, AtomicU64>,
-    system_leap_indicator: Gauge,
-    peer_last_update: Family<PeerLabels, Gauge<f64, AtomicU64>>,
-    peer_poll_interval: Family<PeerLabels, Gauge<f64, AtomicU64>>,
-    peer_poll_interval_exp: Family<PeerLabels, Gauge>,
-    peer_reachability_status: Family<PeerLabels, Gauge>,
-    peer_offset: Family<PeerLabels, Gauge<f64, AtomicU64>>,
-    peer_uncertainty: Family<PeerLabels, Gauge<f64, AtomicU64>>,
-    peer_delay: Family<PeerLabels, Gauge<f64, AtomicU64>>,
-    server_received_packets: Family<ServerLabels, Counter>,
-    server_accepted_packets: Family<ServerLabels, Counter>,
-    server_denied_packets: Family<ServerLabels, Counter>,
-    server_ignored_packets: Family<ServerLabels, Counter>,
-    server_rate_limited_packets: Family<ServerLabels, Counter>,
-    server_response_send_errors: Family<ServerLabels, Counter>,
+impl IntoResponse for ServeError {
+    fn into_response(self) -> axum::response::Response {
+        (
+            StatusCode::INTERNAL_SERVER_ERROR,
+            Json(json!({
+                "error": 500,
+            })),
+        )
+            .into_response()
+    }
 }
 
-impl Metrics {
-    pub fn fill(&self, data: &ObservableState) {
-        let clock = UnixNtpClock::new();
+pub async fn main() -> Result<(), Box<dyn std::error::Error>> {
+    let cli = Cli::parse();
 
-        self.system_poll_interval.set(
-            data.system
-                .time_snapshot
-                .poll_interval
-                .as_duration()
-                .to_seconds(),
-        );
-        self.system_poll_interval_exp
-            .set(data.system.time_snapshot.poll_interval.as_log() as i64);
-        self.system_precision
-            .set(data.system.time_snapshot.precision.to_seconds());
-        self.system_accumulated_steps
-            .set(data.system.time_snapshot.accumulated_steps.to_seconds());
-        self.system_accumulated_steps_threshold.set(
-            data.system
-                .accumulated_steps_threshold
-                .map(|v| v.to_seconds())
-                .unwrap_or(-1.0),
-        );
-        self.system_leap_indicator
-            .set(data.system.time_snapshot.leap_indicator as i64);
+    let config = Config::from_args(cli.config, vec![], vec![]).await;
 
-        for peer in &data.peers {
-            if let ObservablePeerState::Observable {
-                timedata,
-                reachability,
-                poll_interval,
-                address,
-                ..
-            } = peer
-            {
-                let labels = PeerLabels {
-                    address: address.clone(),
-                };
-                self.peer_last_update.get_or_create(&labels).set(
-                    (timedata.last_update
-                        - clock.now().expect("Unable to get current system time"))
-                    .to_seconds(),
-                );
-                self.peer_poll_interval
-                    .get_or_create(&labels)
-                    .set(poll_interval.as_duration().to_seconds());
-                self.peer_poll_interval_exp
-                    .get_or_create(&labels)
-                    .set(poll_interval.as_log() as i64);
-                self.peer_reachability_status
-                    .get_or_create(&labels)
-                    .set(reachability.reachability_score() as i64);
-                self.peer_offset
-                    .get_or_create(&labels)
-                    .set(timedata.offset.to_seconds());
-                self.peer_delay
-                    .get_or_create(&labels)
-                    .set(timedata.delay.to_seconds());
-                self.peer_uncertainty
-                    .get_or_create(&labels)
-                    .set(timedata.uncertainty.to_seconds());
-            }
-        }
-
-        for server in &data.servers {
-            let labels = ServerLabels {
-                listen_address: server.address,
-            };
-
-            self.server_received_packets
-                .get_or_create(&labels)
-                .inner()
-                .store(server.stats.received_packets.get(), Ordering::Relaxed);
-            self.server_accepted_packets
-                .get_or_create(&labels)
-                .inner()
-                .store(server.stats.accepted_packets.get(), Ordering::Relaxed);
-            self.server_denied_packets
-                .get_or_create(&labels)
-                .inner()
-                .store(server.stats.denied_packets.get(), Ordering::Relaxed);
-            self.server_ignored_packets
-                .get_or_create(&labels)
-                .inner()
-                .store(server.stats.ignored_packets.get(), Ordering::Relaxed);
-            self.server_rate_limited_packets
-                .get_or_create(&labels)
-                .inner()
-                .store(server.stats.rate_limited_packets.get(), Ordering::Relaxed);
-            self.server_response_send_errors
-                .get_or_create(&labels)
-                .inner()
-                .store(server.stats.response_send_errors.get(), Ordering::Relaxed);
-        }
+    if let Err(ref e) = config {
+        println!("Warning: Unable to load configuration file: {e}");
     }
 
-    pub fn registry(&self) -> Registry {
-        let mut registry = <Registry>::with_prefix("ntp");
+    let config = config.unwrap_or_default();
 
-        let system = registry.sub_registry_with_prefix("system");
+    let observation_socket_path = match cli.observation_socket {
+        Some(path) => path,
+        None => match config.observe.path {
+            Some(path) => path,
+            None => "/run/ntpd-rs/observe".into(),
+        },
+    };
 
-        system.register_with_unit(
-            "poll_interval",
-            "Time between polls of the system",
-            Unit::Seconds,
-            self.system_poll_interval.clone(),
-        );
-        system.register(
-            "poll_interval",
-            "Exponent of time between poll intervals",
-            self.system_poll_interval_exp.clone(),
-        );
-        system.register_with_unit(
-            "precision",
-            "Precision of the local clock",
-            Unit::Seconds,
-            self.system_precision.clone(),
-        );
-        system.register_with_unit(
-            "accumulated_steps",
-            "Accumulated amount of seconds that the system needed to jump the time",
-            Unit::Seconds,
-            self.system_accumulated_steps.clone(),
-        );
-        system.register_with_unit(
-            "accumulated_steps_threshold",
-            "Threshold for the accumulated step amount at which the NTP daemon will exit (or -1 if no threshold was set)",
-            Unit::Seconds,
-            self.system_accumulated_steps_threshold.clone(),
-        );
-        system.register(
-            "leap_indicator",
-            "Indicates that a leap second will take place",
-            self.system_leap_indicator.clone(),
-        );
+    let app = Router::new()
+        .route(
+            "/metrics",
+            get(|| async {
+                let mut stream = tokio::net::UnixStream::connect(observation_socket_path).await?;
+                let mut msg = Vec::with_capacity(16 * 1024);
+                let output: ObservableState =
+                    ntp_daemon::sockets::read_json(&mut stream, &mut msg).await?;
+                let metrics = Metrics::default();
+                metrics.fill(&output);
+                let registry = metrics.registry();
+                let mut buf = String::new();
+                prometheus_client::encoding::text::encode(&mut buf, &registry)?;
+                Ok::<_, ServeError>((
+                    HeaderMap::from_iter([(
+                        axum::http::header::CONTENT_TYPE,
+                        HeaderValue::from_static("text/plain"),
+                    )]),
+                    buf,
+                ))
+            }),
+        )
+        .route(
+            "/",
+            get(|| async {
+                (
+                    StatusCode::FOUND,
+                    AppendHeaders([(axum::http::header::LOCATION, "/metrics")]),
+                )
+            }),
+        )
+        .fallback(|| async {
+            (
+                StatusCode::NOT_FOUND,
+                Json(json!({
+                    "error": 404,
+                })),
+            )
+        });
 
-        let peer = registry.sub_registry_with_prefix("peer");
+    axum::Server::bind(&cli.listen_socket)
+        .serve(app.into_make_service())
+        .await
+        .unwrap();
 
-        peer.register_with_unit(
-            "uptime",
-            "Time since the peer was started",
-            Unit::Seconds,
-            self.peer_last_update.clone(),
-        );
-
-        peer.register_with_unit(
-            "poll_interval",
-            "Time between polls of the peer",
-            Unit::Seconds,
-            self.peer_poll_interval.clone(),
-        );
-
-        peer.register(
-            "poll_interval",
-            "Exponent of time between polls of the peer",
-            self.peer_poll_interval_exp.clone(),
-        );
-
-        peer.register(
-            "reachability_status",
-            "Number of polls until the upstream server is unreachable, zero if it is",
-            self.peer_reachability_status.clone(),
-        );
-
-        peer.register_with_unit(
-            "offset",
-            "Offset between the upstream server and system time",
-            Unit::Seconds,
-            self.peer_offset.clone(),
-        );
-
-        peer.register_with_unit(
-            "delay",
-            "Current round-trip delay to the upstream server",
-            Unit::Seconds,
-            self.peer_delay.clone(),
-        );
-
-        peer.register_with_unit(
-            "uncertainty",
-            "Estimated error of the clock",
-            Unit::Seconds,
-            self.peer_uncertainty.clone(),
-        );
-
-        let server = registry.sub_registry_with_prefix("server");
-
-        server.register(
-            "received_packets",
-            "Number of incoming received packets",
-            self.server_received_packets.clone(),
-        );
-
-        server.register(
-            "accepted_packets",
-            "Number of packets accepted",
-            self.server_accepted_packets.clone(),
-        );
-
-        server.register(
-            "denied_packets",
-            "Number of denied packets",
-            self.server_denied_packets.clone(),
-        );
-
-        server.register(
-            "ignored_packets",
-            "Number of packets ignored",
-            self.server_ignored_packets.clone(),
-        );
-
-        server.register(
-            "rate_limited_packets",
-            "Number of rate limited packets",
-            self.server_rate_limited_packets.clone(),
-        );
-
-        server.register(
-            "response_send_errors",
-            "Number of packets where there was an error responding",
-            self.server_response_send_errors.clone(),
-        );
-
-        registry
-    }
+    Ok(())
 }
