+++ ntpd/src/bin/ntp-daemon.rs
@@ -1,0 +1,4 @@
+#[tokio::main]
+async fn main() -> Result<(), Box<dyn std::error::Error>> {
+    ntp_daemon::main().await
+}
