+++ ntpd/src/bin/ntp-metrics-exporter.rs
@@ -1,0 +1,4 @@
+#[tokio::main]
+async fn main() -> Result<(), Box<dyn std::error::Error>> {
+    ntp_metrics_exporter::main().await
+}
